requirements.txt
asgiref==3.7.2
defusedxml==0.7.1
diff-match-patch==20230430 
Django==4.2.6
django-filter==23.3        
django-import-export==3.3.1
djangorestframework==3.14.0
et-xmlfile==1.1.0
importlib-metadata==6.8.0
Markdown==3.4.4
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
Pillow==10.0.1
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
sqlparse==0.4.4
tablib==3.5.0
typing_extensions==4.8.0
tzdata==2023.3
whitenoise==6.5.0
xlrd==2.0.1
xlwt==1.3.0
zipp==3.17.0